Added genetic algorithm
Each new generation now consists of children, which have the genes of one parent and undergo light mutation (10%-chance). The results are really interesting. I also added a plot for the minimal score. Now the system is much more dynamic. Depending on the seed, it approaches a state of everyone cooperating (which can hold for more than hundred generations, but it can also happen that they start to betray again)