Fixed a mistake with the Reproduction algorithm
Fixed a grave mistake with the reproduction algorithm (all results before that are wrong) also varied some parameters, added the possibility to mutate the strength of the mutation and made some testruns until cooperation is reached. When Full Cooperation is reached nearly every agent is playing Tit for Tat. But after some generations they lose the ability to defend against defection, so the whole population gets vulnerable against defectors. These can appear through mutation and normally lead to a collapse. In a state of everyone defecting it can take a large amount of time until cooperation starts again. We have to look further into the beginning of cooperation and which criteria have to be fullfilled for a succesful transition